refactor(runtime): move Method to file
